Attempt to clear out substrate deps from kate
