+++ kate/src/lib.rs
@@ -1,9 +1,9 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
-use da_primitives::{BlockLengthColumns, BlockLengthRows};
+use da_types::{BlockLengthColumns, BlockLengthRows};
 #[cfg(feature = "std")]
 pub use dusk_plonk::{commitment_scheme::kzg10::PublicParameters, prelude::BlsScalar};
-use frame_support::sp_runtime::SaturatedConversion;
+use sp_arithmetic::traits::SaturatedConversion;
 #[cfg(feature = "std")]
 use kate_recovery::matrix::Dimensions;
 use static_assertions::const_assert_ne;
