+++ primitives/avail/Cargo.toml
@@ -11,6 +11,7 @@ log = { version = "0.4.8", default-features = false }
 serde = { version = "1.0.121", optional = true, features = ["derive"] }
 serde_json = { version = "1.0", optional = true }
 thiserror-no-std = "2.0.2"
+da-types = { path = "../types", features = ["substrate"] }
 
 # Substrate
 beefy-merkle-tree = { git = "https://github.com/paritytech/substrate.git/", branch = "polkadot-v0.9.37", default-features = false }
